+++ tests/e2e/core-tests/specs/merchant/wp-admin-order-new.test.js
@@ -1,11 +1,10 @@
-/* eslint-disable jest/no-export, jest/no-disabled-tests */
 /**
  * Internal dependencies
  */
 const {
 	merchant,
-	verifyPublishAndTrash,
-	uiUnblocked
+	uiUnblocked,
+	AdminEdit,
 } = require('@woocommerce/e2e-utils');
 const config = require('config');
 const {
@@ -201,11 +200,19 @@ const runCreateOrderTest = () => {
 			await expect(page).toFill('input[name=order_date_minute]', '55');
 
 			// Create order, verify that it was created. Trash order, verify that it was trashed.
-			await verifyPublishAndTrash(
+			const orderEdit = new AdminEdit();
+			await orderEdit.verifyPublish(
 				'.order_actions li .save_order',
 				'#message',
 				'Order updated.',
-				'1 order moved to the Trash.'
+			);
+
+			await expect( page ).toMatchElement( '#select2-order_status-container', { text: 'Processing' } );
+			await expect( page ).toMatchElement(
+				'#woocommerce-order-notes .note_content',
+				{
+					text: 'Order status changed from Pending payment to Processing.',
+				}
 			);
 		});