+++ tests/e2e/core-tests/specs/shopper/front-end-cart.test.js
@@ -24,8 +24,10 @@ const twoProductPrice = singleProductPrice * 2;
 
 const runCartPageTest = () => {
 	describe('Cart page', () => {
+		let productId;
+
 		beforeAll(async () => {
-			await createSimpleProduct();
+			productId = await createSimpleProduct();
 			await withRestApi.resetSettingsGroupToDefault('general');
 			await withRestApi.resetSettingsGroupToDefault('products');
 			await withRestApi.resetSettingsGroupToDefault('tax');
@@ -38,7 +40,7 @@ const runCartPageTest = () => {
 
 		it('should add the product to the cart from the shop page', async () => {
 			await shopper.goToShop();
-			await shopper.addToCartFromShopPage(simpleProductName);
+			await shopper.addToCartFromShopPage( productId );
 
 			await shopper.goToCart();
 			await shopper.productIsInCart(simpleProductName);
@@ -46,7 +48,7 @@ const runCartPageTest = () => {
 
 		it('should increase item qty when "Add to cart" of the same product is clicked', async () => {
 			await shopper.goToShop();
-			await shopper.addToCartFromShopPage(simpleProductName);
+			await shopper.addToCartFromShopPage( productId );
 
 			await shopper.goToCart();
 			await shopper.productIsInCart(simpleProductName, 2);
@@ -63,7 +65,7 @@ const runCartPageTest = () => {
 
 		it('should remove the item from the cart when remove is clicked', async () => {
 			await shopper.goToCart();
-			await shopper.removeFromCart(simpleProductName);
+			await shopper.removeFromCart( productId );
 			await uiUnblocked();
 
 			await expect(page).toMatchElement('.cart-empty', {text: 'Your cart is currently empty.'});
@@ -71,7 +73,7 @@ const runCartPageTest = () => {
 
 		it('should update subtotal in cart totals when adding product to the cart', async () => {
 			await shopper.goToShop();
-			await shopper.addToCartFromShopPage(simpleProductName);
+			await shopper.addToCartFromShopPage( productId );
 
 			await shopper.goToCart();
 			await shopper.productIsInCart(simpleProductName, 1);