+++ tests/e2e/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Unreleased
 
+## Added
+
+- `utils.waitForTimeout( delay )` pause processing for `delay` milliseconds
+- `AdminEdit` class with utility functions for the respective edit screens
+- Update `shopper.addToCartFromShopPage()` and `.removeFromCart()` to accept product Id or Title
+
 # 0.1.6
 
 ## Added